requirements.txt
pycryptodome
sqlcipher3
